% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/myseq_n.R
\name{myseq_n}
\alias{myseq_n}
\title{Get the value from the formula}
\usage{
myseq_n(x, n)
}
\arguments{
\item{x}{numeric}

\item{n}{integer which is >0}
}
\value{
the answer of the recursive sequence
}
\description{
For hw04 function 1, get the answer from the given formula
Xn = Xn−1 + (Xn−3−Xn−2)/n
}
\examples{
myseq_n(x = c(2, 3, 3), n = 3)
myseq_n(x = c(2, 4, 3), n = 4)
myseq_n(x = c(2, 4, 3), n = 5)
myseq_n(x = c(2, 4, 3), n = 6)
myseq_n(x = c(2, 4, 3), n = 7)
}
