% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/myseq_plot.R
\name{myseq_plot}
\alias{myseq_plot}
\title{Produce the plot by given data along using myseq_n() function}
\usage{
myseq_plot(numdf)
}
\arguments{
\item{numdf}{a four col df}
}
\value{
a line plot of the output values for the different values of n
}
\description{
For hw04 function 2, produce the plot by given df.
}
