% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/myseq_n.R
\name{myseq_n}
\alias{myseq_n}
\title{Get the value of the formula}
\usage{
myseq_n(x, n)
}
\arguments{
\item{x}{numeric}

\item{n}{integer which is >0}
}
\value{
the answer of the recursive sequence
}
\description{
For hw04 function 1, get the answer from the given formula.
}
