% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/myseq_plot.R
\name{myseq_plot}
\alias{myseq_plot}
\title{Produce the plot by given data along using myseq_n() function}
\usage{
myseq_plot(numdf)
}
\arguments{
\item{numdf}{a four col df}
}
\value{
a line plot of the output values for the different values of n
}
\description{
For hw04 function 2, produce the plot by given df.
}
\examples{
my_data <- tibble::tribble(
~x, ~y, ~z, ~n,
2,4,3,3,
2,4,3,4,
2,4,3,5,
2,4,3,6,
2,4,3,7,
2,4,3,8,
2,4,3,9,
2,4,3,10,
2,4,3,12)
myseq_plot(numdf = my_data)
}
